requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
mysql-connector-python==9.3.0
requests==2.32.4